refactor: lists
