+++ www/api/v1.3/app/Http/Controllers/SystemController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Http\Controllers\Controller;
+use App\Model\Lists;
+
+class SystemController extends Controller {
+    static public function allLists() {
+        return Lists::getAll();
+    }
+}
