--- www/api/v1.3/public/app/.htaccess
@@ -1,1 +1,0 @@
-Deny from all
