--- www/api/v1.3/public/app/Lib/StringsOps.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Cryptomedic\Lib;
-
-class StringsOps {
-    static function startsWith(string $haystack, string $needle): bool {
-        return substr_compare($haystack, $needle, 0, strlen($needle)) === 0;
-    }
-
-    static function endsWith(string $haystack, string $needle): bool {
-        return substr_compare($haystack, $needle, -strlen($needle)) === 0;
-    }
-}
